Moving the ServiceRouteLoader stuff into the Routing component as an optional feature
